Bug fix (fileService)

+++ server/src/controllers/guilds/editGuildController.js
@@ -2,38 +2,51 @@ import { selectGuildByIdModel } from "../../models/guilds/index.js";
 import { notAuthorizedError } from "../../services/errorService.js";
 import { editGuildModel } from "../../models/guilds/index.js";
 import { deleteFile, saveFile } from "../../services/fileServices.js";
+import validateSchema from "../../schemas/utilities/validateSchema.js";
+import { guildSchema } from "../../schemas/guilds/index.js";
 
 const editGuildController = async (req, res, next) => {
-  try {
-    await validateSchema(editGuildSchema, Object.assign(req.body, req.files));
-    let { name, description } = req.body;
-    const userId = req.user?.id;
-    const guildId = req.params.guildId;
-    const guild = await selectGuildByIdModel(guildId);
+	try {
+		await validateSchema(guildSchema, req.body, req.files);
+		let { name, description } = req.body;
 
-    userId !== guild.owner_id && notAuthorizedError();
-    name = name === guild.name ? null : name;
-    description = description === guild.description ? null : description;
+		console.log("Guild Name", name);
+		const userId = String(req.user?.id);
+		console.log("User ID:", userId);
+		const guildId = req.params.guildId;
+		const guildArray = await selectGuildByIdModel(guildId);
 
-    let fileName;
-    if (req.files) {
-      guild.avatar && (await deleteFile(guild.avatar));
-      const file = req.files;
-      const fileName = await saveFile(file, 150);
-      return fileName;
-    }
+		if (guildArray.length === 0) {
+			return res.status(404).send({
+				status: "error",
+				message: "Hermandad no encontrada",
+			});
+		}
+		const guild = guildArray[0];
+		if (userId !== String(guild.owner_id)) {
+			return notAuthorizedError();
+		}
+		name = name === guild.name ? null : name;
+		description = description === guild.description ? null : description;
 
-    await editGuildModel(guildId, name, description, fileName);
+		let avatar;
+		if (req.files) {
+			guild.avatar && (await deleteFile(guild.avatar));
+			const file = req.files.avatar;
+			avatar = await saveFile(file, 150);
+		}
 
-    res.status(201).send({
-      status: "ok",
-      data: {
-        message: "Hermandad modificada correctamente",
-      },
-    });
-  } catch (err) {
-    next(err);
-  }
+		await editGuildModel(guildId, name, description, avatar);
+
+		res.status(201).send({
+			status: "ok",
+			data: {
+				message: "Hermandad modificada correctamente",
+			},
+		});
+	} catch (err) {
+		next(err);
+	}
 };
 
 export default editGuildController;
